% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{bsw_t0}
\alias{bsw_t0}
\title{Retrieve the time epoch stated in the dimension attribute}
\usage{
bsw_t0(X)
}
\arguments{
\item{X}{ncdf4 object}
}
\value{
POSIXct time origin (possibly different than POSIX epoch origin)
}
\description{
Retrieve the time epoch stated in the dimension attribute
}
