% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsw.R
\name{bsw_tally}
\alias{bsw_tally}
\title{Provides a brief tally/overview of available data}
\usage{
bsw_tally()
}
\value{
a tibble of summary information
}
\description{
Provides a brief tally/overview of available data
}
