% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{bsw_type}
\alias{bsw_type}
\title{Retrieve the bsw type (derived form the path - terrible practice! - but works)}
\usage{
bsw_type(X)
}
\arguments{
\item{X}{ncdf4 object}
}
\value{
charcater of the type
}
\description{
Retrieve the bsw type (derived form the path - terrible practice! - but works)
}
