% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsw.R
\name{bsw_url}
\alias{bsw_url}
\title{Craft a bsw URL for a given date}
\usage{
bsw_url(
  product = c("Aggregation_of_6h_Ocean_Wind", "Aggregation_of_Daily_Ocean_Wind",
    "Aggregation_of_Monthly_Ocean_Wind", "Latest_Climatology_of_Ocean_Winds",
    "Aggregation_of_6h_Ocean_Wind_Stress", "Aggregation_of_Daily_Ocean_Wind_Stress",
    "Aggregation_of_Daily_Ocean_Wind_Stress")[1],
  root = bsw_base_url()
)
}
\arguments{
\item{product}{character, the name of the product}

\item{root}{character, the root URL}
}
\value{
one or more URLs
}
\description{
Craft a bsw URL for a given date
}
