% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{bsw_vars}
\alias{bsw_vars}
\title{Retrieve bsw variables}
\usage{
bsw_vars(X, drop = c("time_run", "time_offset"))
}
\arguments{
\item{X}{ncdf4 object}

\item{drop}{character ietms to not list as variables.  Set to 'none' to drop none.}
}
\value{
character vector
}
\description{
Retrieve bsw variables
}
