% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-method.R
\name{coef,bsw-method}
\alias{coef,bsw-method}
\title{Extracting the estimated model parameters of \code{bsw()}}
\usage{
\S4method{coef}{bsw}(object)
}
\arguments{
\item{object}{An object of class \code{"bsw"}.}
}
\value{
A numeric vector containing the estimated model parameters.
}
\description{
For objects of class \code{"bsw"}, \code{coef()} extracts the estimated model parameters of \code{bsw()}.
}
\author{
Adam Bekhit, Jakob Schöpe
}
