% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constr.R
\name{constr}
\alias{constr}
\title{Setting the linear inequality constraints for \code{bsw()}}
\usage{
constr(x)
}
\arguments{
\item{x}{A model matrix.}
}
\value{
A matrix containing the linear inequality constraints for \code{bsw()}.
}
\description{
\code{constr()} sets the linear inequality constraints for \code{bsw()}.
}
\author{
Adam Bekhit, Jakob Schöpe
}
