% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-method.R
\name{summary,bsw-method}
\alias{summary,bsw-method}
\title{Summarizing the estimated model parameters of \code{bsw()}}
\usage{
\S4method{summary}{bsw}(object)
}
\arguments{
\item{object}{An object of class \code{"bsw"}.}
}
\value{
A list containing the following elements:
\item{coefficients}{A numeric vector containing the estimated model parameters.}
\item{std.err}{A numeric vector containing the estimated standard errors of the model parameters.}
\item{z.value}{A numeric vector containing the estimated z test statistic of the model parameters.}
\item{p.value}{A numeric vector containing the estimated p values of the model parameters.}
}
\description{
For objects of class \code{"bsw"}, \code{summary()} summarizes the estimated model parameters of \code{bsw()}.
}
\author{
Adam Bekhit, Jakob Schöpe
}
