% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsw-package.R
\docType{package}
\name{bsw}
\alias{bsw}
\title{bsw: Simplified access to Blended Sea Winds data from R.}
\description{
Simplified access to Blended Sea Winds data from R.
}
\details{
xyzt-ready. Choose from one of 7 products: 6h (stress or wind), daily (stress or wind), monthly (stress of wind) or climatology (wind).
}
\seealso{
[Blended Sea Winds](https://www.ncei.noaa.gov/products/blended-sea-winds) from NCEI
}
